refactor(tests): Fix test suite failures and improve coverage
This commit addresses multiple test failures encountered during the refactoring of command_handler.py and related components.

Key changes include:

- Resolved asyncio patch-related AttributeErrors.

- Corrected assertions in command_handler memory tests for error string handling.

- Fixed IndexError and linter warnings in port-forward keyboard interrupt tests.

- Simplified port-forward handler tests by directly calling the simplified handler, resolving persistent patching issues related to Config mocking.

- Added missing summary_prompt_func arguments in subcommand calls.

These fixes bring the test suite back to a passing state and improve overall test coverage.
