+++ vibectl/command_handler.py
@@ -8,31 +8,22 @@ Note: All exceptions should propagate to the CLI entry point for centralized err
 handling. Do not print or log user-facing errors here; use logging for diagnostics only.
 """
 
-import asyncio
-import random
-import re
 import subprocess
-import time
 from collections.abc import Callable
-from contextlib import suppress
 
 import click
-import yaml
 from rich.panel import Panel
-from rich.progress import (
-    Progress,
-    SpinnerColumn,
-    TaskID,
-    TextColumn,
-    TimeElapsedColumn,
-)
 from rich.table import Table
 
 from .config import (
     DEFAULT_CONFIG,
     Config,
 )
-from .console import console_manager
+from .k8s_utils import create_kubectl_error, run_kubectl
+from .live_display import (
+    _execute_port_forward_with_live_display,
+    _execute_wait_with_live_display,
+)
 from .logutil import logger as _logger
 from .memory import get_memory, set_memory, update_memory
 from .model_adapter import get_model_adapter
@@ -43,124 +34,17 @@ from .prompt import (
     recovery_prompt,
     wait_resource_prompt,
 )
-from .proxy import (
-    StatsProtocol,
-    TcpProxy,
-    start_proxy_server,
-    stop_proxy_server,
-)
 from .types import Error, OutputFlags, Result, Success
+from .utils import console_manager
 
 logger = _logger
 
 # Export Table for testing
 __all__ = ["Table"]
 
-# Constants for output flags
-# (DEFAULT_MODEL, DEFAULT_SHOW_RAW_OUTPUT, DEFAULT_SHOW_VIBE,
-#  DEFAULT_WARN_NO_OUTPUT, DEFAULT_SHOW_KUBECTL)
-# Use values from config.py's DEFAULT_CONFIG instead
 
 # Initialize output processor
 output_processor = OutputProcessor(max_chars=2000, llm_max_chars=2000)
-
-
-def run_kubectl(
-    cmd: list[str], capture: bool = False, config: Config | None = None
-) -> Result:
-    """Run kubectl command and capture output.
-
-    Args:
-        cmd: List of command arguments
-        capture: Whether to capture and return output
-        config: Optional Config instance to use
-
-    Returns:
-        Success with command output if capture=True, or None otherwise
-        Error with error message on failure
-    """
-    try:
-        # Get a Config instance if not provided
-        cfg = config or Config()
-
-        # Get the kubeconfig path from config
-        kubeconfig = cfg.get("kubeconfig")
-
-        # Build the full command
-        kubectl_cmd = ["kubectl"]
-
-        # Add the command arguments first, to ensure kubeconfig is AFTER the main
-        # command
-        kubectl_cmd.extend(cmd)
-
-        # Add kubeconfig AFTER the main command to avoid errors
-        if kubeconfig:
-            kubectl_cmd.extend(["--kubeconfig", str(kubeconfig)])
-
-        logger.info(f"Running kubectl command: {' '.join(kubectl_cmd)}")
-
-        # Execute the command
-        result = subprocess.run(
-            kubectl_cmd,
-            capture_output=capture,
-            check=False,
-            text=True,
-            encoding="utf-8",
-        )
-
-        # Check for errors
-        if result.returncode != 0:
-            error_message = result.stderr.strip() if capture else "Command failed"
-            if not error_message:
-                error_message = f"Command failed with exit code {result.returncode}"
-            logger.debug(f"kubectl command failed: {error_message}")
-
-            # Create error result, marking kubectl server errors as non-halting
-            # for auto loops
-            return create_kubectl_error(error_message)
-
-        # Return output if capturing
-        if capture:
-            output = result.stdout.strip()
-            logger.debug(f"kubectl command output: {output}")
-            return Success(data=output)
-        return Success()
-    except FileNotFoundError:
-        error_msg = "kubectl not found. Please install it and try again."
-        logger.debug(error_msg)
-        return Error(error=error_msg)
-    except Exception as e:
-        logger.debug(f"Exception running kubectl: {e}", exc_info=True)
-        return Error(error=str(e), exception=e)
-
-
-def create_kubectl_error(
-    error_message: str, exception: Exception | None = None
-) -> Error:
-    """
-    Create an Error object for kubectl failures, marking certain errors as
-    non-halting for auto loops.
-
-    Args:
-        error_message: The error message
-        exception: Optional exception that caused the error
-
-    Returns:
-        Error object with appropriate halt_auto_loop flag set
-    """
-    # For kubectl server errors (like NotFound, Forbidden, etc.),
-    # set halt_auto_loop=False so auto loops can continue
-    if "Error from server" in error_message:
-        return Error(error=error_message, exception=exception, halt_auto_loop=False)
-
-    # For unknown command errors (usually from malformed LLM output),
-    # set halt_auto_loop=False so the auto loop can continue and the LLM can
-    # correct itself
-    if "unknown command" in error_message.lower():
-        return Error(error=error_message, exception=exception, halt_auto_loop=False)
-
-    # For other errors, use the default (halt_auto_loop=True)
-    return Error(error=error_message, exception=exception)
 
 
 def handle_standard_command(
@@ -170,42 +54,55 @@ def handle_standard_command(
     output_flags: OutputFlags,
     summary_prompt_func: Callable[[], str],
 ) -> Result:
-    """Handle a standard kubectl command with both raw and vibe output.
+    """Handle standard kubectl commands like get, describe, logs.
 
     Args:
-        command: The kubectl command to run
-        resource: The resource to act on
-        args: Additional command arguments
-        output_flags: Output configuration flags
-        summary_prompt_func: Function to generate summary prompt
+        command: The kubectl command (get, describe, logs, etc.)
+        resource: The resource type (e.g., pods, deployments)
+        args: Additional arguments for the command
+        output_flags: Flags controlling output format
 
     Returns:
-        Result with Success or Error information
+        Result object containing output or error
     """
-    try:
-        logger.info(f"Handling standard command: {command} {resource} {' '.join(args)}")
+    result = _run_standard_kubectl_command(command, resource, args)
 
-        # Run kubectl and process the result
-        result = _run_standard_kubectl_command(command, resource, args)
-        if isinstance(result, Error):
-            return result
+    if isinstance(result, Error):
+        # Handle API errors specifically if needed
+        if is_api_error(result.error):
+            return create_api_error(result.error, result.exception)
+        # Ensure exception exists before passing
+        if result.exception:
+            return _handle_standard_command_error(
+                command,
+                resource,
+                args,
+                result.exception,
+            )
+        else:
+            # Handle case where Error has no exception (should not happen often)
+            logger.error(
+                f"Command {command} {resource} failed with error but "
+                f"no exception: {result.error}"
+            )
+            return result  # Return the original error
 
-        # Extract output from success result
-        output = result.data
+    output = result.data
 
-        # Handle empty output case
-        if not output:
-            return _handle_empty_output(command, resource, args)
+    # Handle empty output
+    # Ensure output is not None before checking/stripping
+    if output is None or not output.strip():
+        return _handle_empty_output(command, resource, args)
 
-        # Handle the output display based on the configured flags
-        return _handle_standard_command_output(
-            output=output,
-            output_flags=output_flags,
-            summary_prompt_func=summary_prompt_func,
-            command=f"{command} {resource} {' '.join(args)}",
-        )
-    except Exception as e:
-        return _handle_standard_command_error(command, resource, args, e)
+    # Process and display output based on flags
+    # Pass command type to handle_command_output
+    # Output is guaranteed to be a string here
+    return handle_command_output(
+        output,
+        output_flags,
+        summary_prompt_func,
+        command=command,
+    )
 
 
 def _run_standard_kubectl_command(command: str, resource: str, args: tuple) -> Result:
@@ -255,37 +152,6 @@ def _handle_empty_output(command: str, resource: str, args: tuple) -> Result:
     logger.info(f"No output from command: {command} {resource} {' '.join(args)}")
     console_manager.print_processing("Command returned no output")
     return Success(message="Command returned no output")
-
-
-def _handle_standard_command_output(
-    output: str,
-    output_flags: OutputFlags,
-    summary_prompt_func: Callable[[], str],
-    command: str,
-) -> Result:
-    """Process output from a standard command.
-
-    A wrapper around handle_command_output that handles logging.
-
-    Args:
-        output: Command output to process
-        output_flags: Output configuration flags
-        summary_prompt_func: Function to generate summary prompt
-        command: The full command string for logging
-
-    Returns:
-        Result of output processing
-    """
-    output_result = handle_command_output(
-        output=output,
-        output_flags=output_flags,
-        summary_prompt_func=summary_prompt_func,
-        command=command,
-    )
-
-    logger.info(f"Completed standard command: {command}")
-
-    return output_result
 
 
 def _handle_standard_command_error(
@@ -359,92 +225,71 @@ def handle_command_output(
     output: Result | str,
     output_flags: OutputFlags,
     summary_prompt_func: Callable[[], str],
-    max_token_limit: int = 10000,
-    truncation_ratio: int = 3,
     command: str | None = None,
 ) -> Result:
-    """Handle displaying command output in both raw and vibe formats.
+    """Processes and displays command output based on flags.
 
     Args:
-        output: The command output to display (can be a string or a Result object)
-        output_flags: Configuration for output display
-        summary_prompt_func: Function returning the prompt template for summarizing
-        max_token_limit: Maximum number of tokens for the prompt
-        truncation_ratio: Ratio for truncating the output
-        command: Optional command string that generated the output
+        output: The command output string or a Result object.
+        output_flags: Flags controlling the output format.
+        max_token_limit: Max token limit for LLM input.
+        truncation_ratio: Ratio for truncating long output.
+        command: The original kubectl command type (e.g., get, describe).
 
     Returns:
-        Result with Success containing vibe output or Error with error information
+        Result object containing the processed output or original error.
     """
-    logger.debug(f"Handling command output for: {command}")
+    _check_output_visibility(output_flags)
 
-    try:
-        # Extract data from Result objects if needed
-        if isinstance(output, Success):
-            output_data = output.data
-        elif isinstance(output, Error):
-            return output  # Pass through error results
-        else:
-            output_data = output  # Already a string
+    if isinstance(output, Error):
+        # If input is an Error, display it and return
+        console_manager.print_error(output.error)
+        return output
 
-        # Ensure we have string data to work with
-        if not isinstance(output_data, str):
-            logger.warning(
-                f"Expected string output but got {type(output_data).__name__}"
-            )
-            # Convert to string if possible
-            output_data = "" if output_data is None else str(output_data)
+    # Extract string output from Success object if needed
+    output_str = output.data if isinstance(output, Success) else output
 
-        # Display command if needed
-        _display_kubectl_command(output_flags, command)
+    _display_kubectl_command(output_flags, command)
+    # Ensure output_str is not None before displaying
+    if output_str is not None:
+        _display_raw_output(output_flags, output_str)
 
-        # Check and warn if no output will be displayed
-        _check_output_visibility(output_flags)
+    # Determine if Vibe processing is needed
+    if output_flags.show_vibe:
+        summary_prompt_str = summary_prompt_func()
 
-        # Show raw output if needed
-        _display_raw_output(output_flags, output_data)
-
-        # Process vibe output if needed
-        if output_flags.show_vibe:
-            vibe_result = _process_vibe_output(
-                output_data,
-                output_flags,
-                summary_prompt_func,
-                max_token_limit,
-                truncation_ratio,
-                command,
-            )
-
-            # If we got an error from vibe processing, return that
-            if isinstance(vibe_result, Error):
+        # Pass the formatted string to _process_vibe_output
+        # Ensure output_str is not None before processing
+        if output_str is not None:
+            try:
+                vibe_result = _process_vibe_output(
+                    output_str,
+                    output_flags,
+                    summary_prompt_str=summary_prompt_str,  # Pass formatted string
+                    command=command,
+                )
+                # Directly return the result from vibe processing
                 return vibe_result
-
-            # Otherwise, extract the vibe output
-            vibe_output = vibe_result.data
-
-            # Display the output if there was any
-            if vibe_output and output_flags.show_raw:
-                console_manager.console.print()
-
-            if vibe_output:
-                _display_vibe_output(vibe_output)
+            except Exception as e:
+                logger.error(f"Error during Vibe processing: {e}", exc_info=True)
+                console_manager.print_error(f"Error processing Vibe output: {e}")
+                # Return an Error object after handling the exception
+                error_str = str(e)
+                if is_api_error(error_str):
+                    return create_api_error(error_str, exception=e)
+                else:
+                    return Error(error=error_str, exception=e)
         else:
-            vibe_output = ""
+            # Handle case where output_str is None but Vibe was requested
+            logger.warning("Cannot process Vibe output because input output was None.")
+            return Error(
+                error="Input command output was None, cannot generate Vibe summary."
+            )
 
-        # Return success with whatever vibe output we got
-        return Success(message="Command completed successfully", data=vibe_output)
-
-    except Exception as e:
-        logger.error(f"Error in vibe output processing: {e}", exc_info=True)
-        error_message = f"Error processing output: {e}"
-
-        # Check if this is an API error, which shouldn't halt auto loops
-        if is_api_error(str(e)):
-            logger.info("API error detected, marking as non-halting for auto loops")
-            return create_api_error(error_message, e)
-
-        # For non-API errors, use the default behavior (halt_auto_loop=True)
-        return Error(error=error_message, exception=e)
+    else:
+        # If only raw output is requested, return it as Success
+        # Ensure output_str is not None before returning
+        return Success(message=output_str if output_str is not None else "")
 
 
 def _display_kubectl_command(output_flags: OutputFlags, command: str | None) -> None:
@@ -504,108 +349,81 @@ def _display_raw_output(output_flags: OutputFlags, output: str) -> None:
 def _process_vibe_output(
     output: str,
     output_flags: OutputFlags,
-    summary_prompt_func: Callable[[], str],
-    max_token_limit: int = 10000,
-    truncation_ratio: int = 3,
+    summary_prompt_str: str,
     command: str | None = None,
 ) -> Result:
-    """Process command output to generate a vibe summary.
+    """Processes output using Vibe LLM for summary.
 
     Args:
-        output: The command output to process
-        output_flags: Output configuration flags
-        summary_prompt_func: Function to generate summary prompt
-        max_token_limit: Maximum number of tokens for the prompt
-        truncation_ratio: Ratio for truncating the output
-        command: Optional command string that generated the output
+        output: The raw command output string.
+        output_flags: Flags controlling output format.
+        summary_prompt_str: The formatted prompt string for the LLM.
+
+        command: The original kubectl command type.
 
     Returns:
-        Result with Success containing vibe output or Error with error information
+        Result object with Vibe summary or an Error.
     """
-    logger.debug("Processing output for vibe summary.")
+    # Truncate output if necessary
+    processed_output = output_processor.process_auto(output).truncated
 
-    # Process output to avoid token limits
-    truncation_result = output_processor.process_auto(output)
-    processed_output = truncation_result.truncated
-
-    # Show truncation warning if needed
-    # Simplification: Check if truncated content is different from original
-    if truncation_result.original != truncation_result.truncated:
-        logger.warning("Output was truncated for processing.")
-        console_manager.print_truncation_warning()
-
+    # Get LLM summary
     try:
-        # Get summary from LLM
         vibe_output = _get_llm_summary(
-            processed_output, output_flags.model_name, summary_prompt_func(), command
+            processed_output,
+            output_flags.model_name,
+            summary_prompt_str,  # Pass formatted string
         )
 
-        # Update memory if we have a command, regardless of vibe output
-        if command:
-            update_memory(command, output, vibe_output, output_flags.model_name)
-
-        # Check for empty response
-        if not vibe_output:
-            logger.info("Vibe output is empty.")
-            console_manager.print_empty_output_message()
-            return Success(
-                message="Command completed successfully, but vibe output is empty.",
-                data="",
-            )
-
-        # Check for error response
+        # Check if the LLM returned an error string
         if vibe_output.startswith("ERROR:"):
-            error_message = vibe_output[7:].strip()  # Remove "ERROR: " prefix
-            logger.error(f"Vibe model returned error: {error_message}")
-            console_manager.print_error(error_message)
-            return Error(error=error_message)
+            error_message = vibe_output[7:].strip()
+            logger.error(f"LLM summary error: {error_message}")
+            console_manager.print_error(vibe_output)  # Display the full ERROR: string
+            # Check if it's an API error to set halt_auto_loop correctly
+            if is_api_error(error_message):
+                # Pass the error message without the ERROR: prefix
+                return create_api_error(error_message)
+            else:
+                return Error(error=error_message)
 
-        # Return the vibe output
-        return Success(message="Vibe output processed successfully", data=vibe_output)
+        _display_vibe_output(vibe_output)
 
+        # Update memory only if Vibe summary succeeded (and wasn't an error string)
+        update_memory(
+            command=command or "Unknown",
+            command_output=output,  # Store original full output in memory
+            vibe_output=vibe_output,
+            model_name=output_flags.model_name,
+        )
+        return Success(message=vibe_output)
     except Exception as e:
-        error_message = f"Error getting LLM summary: {e}"
-
-        # Check if this is an API error, which shouldn't halt auto loops
-        if is_api_error(str(e)):
-            logger.info(
-                "API error in LLM summary, marking as non-halting for auto loops"
-            )
-            return create_api_error(error_message, e)
-
-        # For other errors, use the default behavior
-        return Error(error=error_message, exception=e)
+        logger.error(f"Error getting Vibe summary: {e}", exc_info=True)
+        console_manager.print_error(f"Error processing Vibe output: {e}")
+        return Error(error=str(e), exception=e)
 
 
 def _get_llm_summary(
     processed_output: str,
     model_name: str,
-    summary_prompt: str,
-    command: str | None = None,
+    summary_prompt_str: str,
 ) -> str:
-    """Get a summary from the LLM.
+    """Gets the LLM summary for the processed output.
 
     Args:
-        processed_output: Processed command output
-        model_name: Name of the model to use
-        summary_prompt: Prompt template for summarizing
-        command: Optional command string to include in the prompt
+        processed_output: The processed (potentially truncated) output.
+        model_name: Name of the LLM model to use.
+        summary_prompt_str: The formatted prompt string for the LLM.
+        command: The original kubectl command type.
 
     Returns:
-        Summary text from the LLM
+        The summary generated by the LLM.
     """
     model_adapter = get_model_adapter()
     model = model_adapter.get_model(model_name)
-
-    # Format the prompt correctly based on available parameters
-    prompt = (
-        summary_prompt.format(output=processed_output, command=command)
-        if command
-        else summary_prompt.format(output=processed_output)
-    )
-
-    logger.debug(f"Sending prompt to model: {prompt[:100]}...")
-    return model_adapter.execute(model, prompt)
+    # Format the prompt string with the output
+    final_prompt = summary_prompt_str.format(output=processed_output)
+    return model_adapter.execute(model, final_prompt)
 
 
 def _display_vibe_output(vibe_output: str) -> None:
@@ -654,10 +472,17 @@ def handle_vibe_request(
         model_adapter = get_model_adapter()
         model = model_adapter.get_model(output_flags.model_name)
 
-        # Format the prompt for the LLM
-        formatted_prompt = _format_vibe_prompt(
-            plan_prompt, request, command, memory_context
-        )
+        # Format the prompt for the LLM - Directly format plan_prompt
+        # Remove call to _format_vibe_prompt
+        try:
+            formatted_prompt = plan_prompt.format(request=request)
+        except KeyError as e:
+            logger.warning(
+                f"Prompt formatting failed for request '{request}' "
+                f"(Format error: {e}). Using fallback formatting method."
+            )
+            # Fallback if prompt doesn't use {request} key
+            formatted_prompt = plan_prompt
 
         # Get the kubectl command from the LLM
         kubectl_cmd = model_adapter.execute(model, formatted_prompt)
@@ -694,7 +519,7 @@ def handle_vibe_request(
             logger.info("Command execution failed, getting recovery suggestions")
             # Get recovery suggestions from model
             try:
-                # Format the recovery prompt
+                # Format the recovery prompt (only needs command and error)
                 recovery_prompt_text = recovery_prompt(
                     command=f"{command} {kubectl_cmd}",
                     error=result.error,
@@ -727,80 +552,6 @@ def handle_vibe_request(
     except Exception as e:
         logger.error(f"Error in vibe request processing: {e}", exc_info=True)
         return Error(error=f"Error processing vibe request: {e}", exception=e)
-
-
-def _format_vibe_prompt(
-    plan_prompt: str, request: str, command: str, memory_context: str = ""
-) -> str:
-    """Format the plan prompt for vibe requests, handling different format styles.
-
-    Args:
-        plan_prompt: The prompt template
-        request: The user's natural language request
-        command: The command type (get, describe, etc.)
-        memory_context: Optional memory context to include
-
-    Returns:
-        Formatted prompt ready to send to the LLM
-    """
-    # Prepare the format parameters
-    format_params = {"request": request, "command": command}
-
-    # Always provide memory_context in format_params, even if empty
-    # This prevents KeyError when the prompt contains {memory_context}
-    format_params["memory_context"] = memory_context or ""
-
-    try:
-        # First, check if there are any positional format specifiers in the prompt
-        import re
-
-        positional_formats = re.findall(r"{(\d+)}", plan_prompt)
-
-        if positional_formats:
-            # If positional formats exist, use string replacement for all parameters
-            logger.info(
-                "Detected positional format specifiers in prompt, "
-                "using string replacement"
-            )
-            formatted_prompt = plan_prompt
-
-            # Replace all named parameters
-            formatted_prompt = formatted_prompt.replace(
-                "{memory_context}", memory_context or ""
-            )
-            formatted_prompt = formatted_prompt.replace("{request}", request)
-            formatted_prompt = formatted_prompt.replace("{command}", command)
-
-            # Then handle any remaining positional parameters by replacing
-            # them with empty strings
-            for pos in positional_formats:
-                formatted_prompt = formatted_prompt.replace(f"{{{pos}}}", "")
-        else:
-            # No positional formats, use normal keyword formatting
-            formatted_prompt = plan_prompt.format(**format_params)
-
-    except (KeyError, IndexError) as e:
-        # Fallback to string replacement as a last resort
-        logger.warning(
-            f"Format error ({e}) in prompt. Using fallback formatting method."
-        )
-        # Use string replacement as a fallback to avoid format conflicts
-        formatted_prompt = plan_prompt
-        formatted_prompt = formatted_prompt.replace(
-            "{memory_context}", memory_context or ""
-        )
-        formatted_prompt = formatted_prompt.replace("{request}", request).replace(
-            "{command}", command
-        )
-
-        # Replace any remaining format specifiers with empty strings
-        import re
-
-        formatted_prompt = re.sub(r"{(\d+)}", "", formatted_prompt)
-        # Also replace any remaining {name} format specifiers
-        formatted_prompt = re.sub(r"{[a-zA-Z0-9_]+}", "", formatted_prompt)
-
-    return formatted_prompt
 
 
 def _handle_planning_error(
@@ -920,7 +671,14 @@ def _process_and_execute_kubectl_command(
         # Get the resource and args from the parsed command
         resource = args[1] if len(args) > 1 else ""
         wait_args = tuple(args[2:]) if len(args) > 2 else ()
-        return handle_wait_with_live_display(resource, wait_args, modified_output_flags)
+
+        # TODO: refactor for cleaner control and explicit passing of prompt...
+        return handle_wait_with_live_display(
+            resource,
+            wait_args,
+            modified_output_flags,
+            summary_prompt_func=wait_resource_prompt,
+        )
 
     # Handle port-forward with live display
     if (
@@ -931,8 +689,13 @@ def _process_and_execute_kubectl_command(
     ):
         resource = args[1] if len(args) > 1 else ""
         port_args = tuple(args[2:]) if len(args) > 2 else ()
+
+        # TODO: refactor for cleaner control and explicit passing of prompt...
         return handle_port_forward_with_live_display(
-            resource, port_args, modified_output_flags
+            resource,
+            port_args,
+            modified_output_flags,
+            summary_prompt_func=port_forward_prompt,
         )
 
     # Execute the standard kubectl command
@@ -1089,19 +852,22 @@ def _handle_fuzzy_memory_update(option: str, model_name: str) -> Result:
     update_text = click.prompt("Memory update")
 
     # Update memory with the provided text
-    from vibectl.config import Config
-
     try:
         # Get the model name from config if not specified
         cfg = Config()
-        current_memory = get_memory()
+        current_memory = get_memory(cfg)  # Pass cfg
 
         # Get the model
-        model_adapter = get_model_adapter()
+        model_adapter = get_model_adapter(cfg)  # Pass cfg
         model = model_adapter.get_model(model_name)
 
         # Create a prompt for the fuzzy memory update
-        prompt = memory_fuzzy_update_prompt(current_memory, update_text, cfg)
+        # Pass context arguments explicitly to memory_fuzzy_update_prompt if required
+        # Assuming memory_fuzzy_update_prompt handles context internally via config
+        prompt = memory_fuzzy_update_prompt(
+            current_memory=current_memory,
+            update_text=update_text,
+        )
 
         # Get the response
         console_manager.print_processing("Updating memory...")
@@ -1309,7 +1075,6 @@ def _execute_command_with_complex_args(args: list[str]) -> Result:
     Returns:
         Result with Success containing command output or Error with error information
     """
-    import subprocess
 
     try:
         # Build the full command to preserve argument structure
@@ -1348,7 +1113,6 @@ def _execute_yaml_command(args: list[str], yaml_content: str) -> Result:
         Result with Success containing command output or Error with error information
     """
     import re
-    import subprocess
     import tempfile
     from subprocess import TimeoutExpired
 
@@ -1525,20 +1289,35 @@ def configure_output_flags(
     )
 
 
+def parse_kubectl_command(command_string: str) -> tuple[str, list[str]]:
+    """Parses a kubectl command string into command and arguments."""
+    # Split the command string into command and arguments
+    parts = command_string.split(maxsplit=1)
+    if len(parts) > 1:
+        command = parts[0]
+        args = parts[1].split()
+    else:
+        command = parts[0]
+        args = []
+    return command, args
+
+
+# Wrapper for wait command live display
 def handle_wait_with_live_display(
     resource: str,
     args: tuple[str, ...],
     output_flags: OutputFlags,
+    summary_prompt_func: Callable[[], str],
 ) -> Result:
-    """Handle wait command with a live spinner and elapsed time display.
+    """Handles `kubectl wait` by preparing args and calling the live display worker.
 
     Args:
-        resource: Resource to wait for
-        args: Command line arguments
-        output_flags: Output configuration flags
+        resource: The resource type (e.g., pod, deployment).
+        args: Command arguments including resource name and conditions.
+        output_flags: Flags controlling output format.
 
     Returns:
-        Result with Success containing wait output or Error with error information
+        Result from the live display worker function.
     """
     # Extract the condition from args for display
     condition = "condition"
@@ -1550,248 +1329,47 @@ def handle_wait_with_live_display(
     # Create the command for display
     display_text = f"Waiting for {resource} to meet {condition}"
 
-    # Track start time to calculate total duration
-    start_time = time.time()
+    # Call the worker function in live_display.py
+    wait_result = _execute_wait_with_live_display(
+        resource=resource,
+        args=args,
+        output_flags=output_flags,
+        condition=condition,
+        display_text=display_text,
+    )
 
-    # This is our async function to run the kubectl wait command
-    async def async_run_wait_command() -> Result:
-        """Run kubectl wait command asynchronously."""
-        # Build command list
-        cmd_args = ["wait", resource]
-        if args:
-            cmd_args.extend(args)
-
-        # Execute the command in a separate thread to avoid blocking the event loop
-        # We use asyncio.to_thread to run the blocking kubectl call in a thread pool
-        return await asyncio.to_thread(run_kubectl, cmd_args, capture=True)
-
-    # Create a coroutine to update the progress display continuously
-    async def update_progress(task_id: TaskID, progress: Progress) -> None:
-        """Update the progress display regularly."""
-        try:
-            # Keep updating at a frequent interval until cancelled
-            while True:
-                progress.update(task_id)
-                # Very small sleep interval for smoother animation
-                # (20-30 updates per second)
-                await asyncio.sleep(0.03)
-        except asyncio.CancelledError:
-            # Handle cancellation gracefully by doing a final update
-            progress.update(task_id)
-            return
-
-    # Create a more visually appealing progress display
-    with Progress(
-        SpinnerColumn(),
-        TimeElapsedColumn(),
-        TextColumn("[bold blue]{task.description}"),
-        console=console_manager.console,
-        transient=True,
-        refresh_per_second=30,  # Higher refresh rate for smoother animation
-    ) as progress:
-        # Add a wait task
-        task_id = progress.add_task(description=display_text, total=None)
-
-        # Define the async main routine that coordinates the wait operation
-        async def main() -> Result:
-            """Main async routine that runs the wait command and updates progress."""
-            # Start updating the progress display in a separate task
-            progress_task = asyncio.create_task(update_progress(task_id, progress))
-
-            # Force at least one update to ensure spinner visibility
-            await asyncio.sleep(0.1)
-
-            try:
-                # Run the wait command
-                result = await async_run_wait_command()
-
-                # Give the progress display time to show completion
-                # (avoids abrupt disappearance)
-                await asyncio.sleep(0.5)
-
-                # Cancel the progress update task
-                if not progress_task.done():
-                    progress_task.cancel()
-                    # Wait for the task to actually cancel
-                    with suppress(asyncio.TimeoutError, asyncio.CancelledError):
-                        await asyncio.wait_for(progress_task, timeout=0.5)
-
-                return result
-            except Exception as e:
-                # Ensure we cancel the progress task on errors
-                if not progress_task.done():
-                    progress_task.cancel()
-                    with suppress(asyncio.TimeoutError, asyncio.CancelledError):
-                        await asyncio.wait_for(progress_task, timeout=0.5)
-
-                # Return an error result
-                return Error(error=str(e), exception=e)
-
-        # Set up loop and run the async code
-        result = None
-        created_new_loop = False
-        loop = None
-        wait_success = False  # Track if wait completed successfully
-
-        try:
-            # Get or create an event loop in a resilient way
-            try:
-                loop = asyncio.get_event_loop()
-                if loop.is_running():
-                    # If we're in a running loop context, create a new one
-                    loop = asyncio.new_event_loop()
-                    asyncio.set_event_loop(loop)
-                    created_new_loop = True
-            except RuntimeError:
-                # If we can't get a loop, create a new one
-                loop = asyncio.new_event_loop()
-                asyncio.set_event_loop(loop)
-                created_new_loop = True
-
-            # Run our main coroutine in the event loop
-            result = loop.run_until_complete(main())
-            wait_success = isinstance(
-                result, Success
-            )  # Success if we got a Success result
-
-        except asyncio.CancelledError:
-            # Handle user interrupts (like Ctrl+C)
-            console_manager.print_note("Wait operation cancelled")
-            return Error(error="Wait operation cancelled by user")
-
-        finally:
-            # Clean up the progress display
-            progress.stop()
-
-            # If we created a new loop, close it to prevent asyncio warnings
-            if created_new_loop and loop is not None:
-                loop.close()
-
-    # Calculate elapsed time regardless of output
-    elapsed_time = time.time() - start_time
-
-    # Handle the command output if any
-    if wait_success and isinstance(result, Success):
-        # Display success message with duration
-        console_manager.console.print(
-            f"[bold green]✓[/] Wait completed in [bold]{elapsed_time:.2f}s[/]"
-        )
-
-        # Add a small visual separator before the output
-        if output_flags.show_raw or output_flags.show_vibe:
-            console_manager.console.print()
-
-        output_result = handle_command_output(
-            output=result.data or "",
-            output_flags=output_flags,
-            summary_prompt_func=wait_resource_prompt,
-            command=f"wait {resource} {' '.join(args)}",
-        )
-        return output_result
-    elif wait_success:
-        # If wait completed successfully but there's no output to display
-        success_message = (
-            f"[bold green]✓[/] {resource} now meets condition '[bold]{condition}[/]' "
-            f"(completed in [bold]{elapsed_time:.2f}s[/])"
-        )
-        console_manager.safe_print(console_manager.console, success_message)
-
-        # Add a small note if no output will be shown
-        if not output_flags.show_raw and not output_flags.show_vibe:
-            message = (
-                "\nNo output display enabled. Use --show-raw-output or "
-                "--show-vibe to see details."
-            )
-            console_manager.console.print(message)
-
-        return Success(
-            message=(
-                f"{resource} now meets condition '{condition}' "
-                f"(completed in {elapsed_time:.2f}s)"
-            ),
-        )
-    else:
-        # If there was an issue but we didn't raise an exception
-        if isinstance(result, Error):
-            message = (
-                f"[bold red]✗[/] Wait operation failed after "
-                f"[bold]{elapsed_time:.2f}s[/]: {result.error}"
-            )
-            console_manager.safe_print(console_manager.console, message)
-            return result
-        else:
-            message = (
-                f"[bold yellow]![/] Wait operation completed with no result "
-                f"after [bold]{elapsed_time:.2f}s[/]"
-            )
-            console_manager.console.print(message)
-            return Error(
-                error=(
-                    f"Wait operation completed with no result after {elapsed_time:.2f}s"
-                )
-            )
-
-
-class ConnectionStats(StatsProtocol):
-    """Track connection statistics for port-forward sessions."""
-
-    def __init__(self) -> None:
-        """Initialize connection statistics."""
-        self.current_status = "Connecting"  # Current connection status
-        self.connections_attempted = 0  # Number of connection attempts
-        self.successful_connections = 0  # Number of successful connections
-        self.bytes_sent = 0  # Bytes sent through connection
-        self.bytes_received = 0  # Bytes received through connection
-        self.elapsed_connected_time = 0.0  # Time in seconds connection was active
-        self.traffic_monitoring_enabled = False  # Whether traffic stats are available
-        self.using_proxy = False  # Whether connection is going through proxy
-        self.error_messages: list[str] = []  # List of error messages encountered
-        self._last_activity_time = time.time()  # Timestamp of last activity
-
-    @property
-    def last_activity(self) -> float:
-        """Get the timestamp of the last activity."""
-        return self._last_activity_time
-
-    @last_activity.setter
-    def last_activity(self, value: float) -> None:
-        """Set the timestamp of the last activity."""
-        self._last_activity_time = value
-
-
-def has_port_mapping(port_mapping: str) -> bool:
-    """Check if a valid port mapping is provided.
-
-    Args:
-        port_mapping: The port mapping string to check
-
-    Returns:
-        True if a valid port mapping with format "local:remote" is provided
-    """
-    return ":" in port_mapping and all(
-        part.isdigit() for part in port_mapping.split(":")
+    # Process the result from the worker using handle_command_output
+    # Create the command string for context
+    command_str = f"wait {resource} {' '.join(args)}"
+    return handle_command_output(
+        output=wait_result,  # Pass the Result object directly
+        output_flags=output_flags,
+        summary_prompt_func=summary_prompt_func,
+        command=command_str,
     )
 
 
+# Wrapper for port-forward command live display
 def handle_port_forward_with_live_display(
     resource: str,
     args: tuple[str, ...],
     output_flags: OutputFlags,
+    summary_prompt_func: Callable[[], str],
 ) -> Result:
-    """Handle port-forward command with a live display showing connection status
-    and ports.
+    """Handles `kubectl port-forward` by preparing args and invoking live display.
 
     Args:
-        resource: Resource to forward ports for
-        args: Command line arguments including port specifications
-        output_flags: Output configuration flags
+        resource: The resource type (e.g., pod, service).
+        args: Command arguments including resource name and port mappings.
+        output_flags: Flags controlling output format.
 
     Returns:
-        Result with Success containing port-forward information or Error on failure
+        Result from the live display worker function.
     """
     # Extract port mapping from args for display
     port_mapping = "port"
     for arg in args:
+        # Simple check for port mapping format (e.g., 8080:80)
         if ":" in arg and all(part.isdigit() for part in arg.split(":")):
             port_mapping = arg
             break
@@ -1807,436 +1385,14 @@ def handle_port_forward_with_live_display(
         f"to localhost:[bold]{local_port}[/]"
     )
 
-    # Track start time for elapsed time display
-    start_time = time.time()
-
-    # Create a stats object to track connection information
-    stats = ConnectionStats()
-
-    # Check if traffic monitoring is enabled via intermediate port range
-    cfg = Config()
-    intermediate_port_range = cfg.get("intermediate_port_range")
-    use_proxy = False
-    proxy_port = None
-
-    # Check if a port mapping was provided (required for proxy)
-    has_valid_port_mapping = has_port_mapping(port_mapping)
-
-    if intermediate_port_range and has_valid_port_mapping:
-        try:
-            # Parse the port range
-            min_port, max_port = map(int, intermediate_port_range.split("-"))
-
-            # Get a random port in the range
-            proxy_port = random.randint(min_port, max_port)
-
-            # Enable proxy mode
-            use_proxy = True
-            stats.using_proxy = True
-            stats.traffic_monitoring_enabled = True
-
-            console_manager.print_note(
-                f"Traffic monitoring enabled via proxy on port {proxy_port}"
-            )
-        except (ValueError, AttributeError) as e:
-            console_manager.print_error(
-                f"Invalid intermediate_port_range format: {intermediate_port_range}. "
-                f"Expected format: 'min-max'. Error: {e}"
-            )
-            use_proxy = False
-            return Error(
-                error=(
-                    f"Invalid intermediate_port_range format: "
-                    f"{intermediate_port_range}. Expected format: 'min-max'."
-                ),
-                exception=e,
-            )
-    elif (
-        not intermediate_port_range
-        and has_valid_port_mapping
-        and output_flags.warn_no_proxy
-    ):
-        # Show warning about missing proxy configuration when port mapping is provided
-        console_manager.print_no_proxy_warning()
-
-    # Create a subprocess to run kubectl port-forward
-    # We'll use asyncio to manage this process and update the display
-    async def run_port_forward() -> asyncio.subprocess.Process:
-        """Run the port-forward command and capture output."""
-        # Build command list
-        cmd_args = ["port-forward", resource]
-
-        # Make sure we have valid args - check for resource pattern first
-        args_list = list(args)
-
-        # If using proxy, modify the port mapping argument to use proxy_port
-        if use_proxy and proxy_port is not None:
-            # Find and replace the port mapping argument
-            for i, arg in enumerate(args_list):
-                if ":" in arg and all(part.isdigit() for part in arg.split(":")):
-                    # Replace with proxy port:remote port
-                    args_list[i] = f"{proxy_port}:{remote_port}"
-                    break
-
-        # Add remaining arguments
-        if args_list:
-            cmd_args.extend(args_list)
-
-        # Full kubectl command
-        kubectl_cmd = ["kubectl"]
-
-        # Add kubeconfig if set
-        kubeconfig = cfg.get("kubeconfig")
-        if kubeconfig:
-            kubectl_cmd.extend(["--kubeconfig", str(kubeconfig)])
-
-        # Add the port-forward command args
-        kubectl_cmd.extend(cmd_args)
-
-        # Create a process to run kubectl port-forward
-        # This process will keep running until cancelled
-        process = await asyncio.create_subprocess_exec(
-            *kubectl_cmd,
-            stdout=asyncio.subprocess.PIPE,
-            stderr=asyncio.subprocess.PIPE,
-        )
-
-        # Increment connection attempts counter
-        stats.connections_attempted += 1
-
-        # Return reference to the process
-        return process
-
-    # Update the progress display with connection status
-    async def update_progress(
-        task_id: TaskID,
-        progress: Progress,
-        process: asyncio.subprocess.Process,
-        proxy: TcpProxy | None = None,
-    ) -> None:
-        """Update the progress display with connection status and data."""
-        connected = False
-        connection_start_time = None
-
-        try:
-            # Keep updating until cancelled
-            while True:
-                # Check if process has output ready
-                if process.stdout:
-                    line = await process.stdout.readline()
-                    if line:
-                        # Got output, update connection status
-                        line_str = line.decode("utf-8").strip()
-                        if "Forwarding from" in line_str:
-                            connected = True
-                            stats.current_status = "Connected"
-                            stats.successful_connections += 1
-                            if connection_start_time is None:
-                                connection_start_time = time.time()
-
-                            # Attempt to parse traffic information if available
-                            if "traffic" in line_str.lower():
-                                stats.traffic_monitoring_enabled = True
-                                # Extract bytes sent/received if available
-                                # Parsing depends on the output format
-                                if "sent" in line_str.lower():
-                                    sent_match = re.search(
-                                        r"sent (\d+)", line_str.lower()
-                                    )
-                                    if sent_match:
-                                        stats.bytes_sent += int(sent_match.group(1))
-                                if "received" in line_str.lower():
-                                    received_match = re.search(
-                                        r"received (\d+)", line_str.lower()
-                                    )
-                                    if received_match:
-                                        stats.bytes_received += int(
-                                            received_match.group(1)
-                                        )
-
-                # Update stats from proxy if enabled
-                if proxy and connected:
-                    # Update stats from the proxy server
-                    stats.bytes_sent = proxy.stats.bytes_sent
-                    stats.bytes_received = proxy.stats.bytes_received
-                    stats.traffic_monitoring_enabled = True
-
-                # Update connection time if connected
-                if connected and connection_start_time is not None:
-                    stats.elapsed_connected_time = time.time() - connection_start_time
-
-                # Update the description based on connection status
-                if connected:
-                    if proxy:
-                        # Show traffic stats in the description when using proxy
-                        bytes_sent = stats.bytes_sent
-                        bytes_received = stats.bytes_received
-                        progress.update(
-                            task_id,
-                            description=(
-                                f"{display_text} - [green]Connected[/green] "
-                                f"([cyan]↑{bytes_sent}B[/] "
-                                f"[magenta]↓{bytes_received}B[/])"
-                            ),
-                        )
-                    else:
-                        progress.update(
-                            task_id,
-                            description=f"{display_text} - [green]Connected[/green]",
-                        )
-                else:
-                    # Check if the process is still running
-                    if process.returncode is not None:
-                        stats.current_status = "Disconnected"
-                        progress.update(
-                            task_id,
-                            description=f"{display_text} - [red]Disconnected[/red]",
-                        )
-                        break
-
-                    # Still establishing connection
-                    progress.update(
-                        task_id,
-                        description=f"{display_text} - Connecting...",
-                    )
-
-                # Small sleep for smooth updates
-                await asyncio.sleep(0.1)
-
-        except asyncio.CancelledError:
-            # Final update before cancellation
-            stats.current_status = "Cancelled"
-            progress.update(
-                task_id,
-                description=f"{display_text} - [yellow]Cancelled[/yellow]",
-            )
-
-    # Create progress display
-    with Progress(
-        SpinnerColumn(),
-        TimeElapsedColumn(),
-        TextColumn("{task.description}"),
-        console=console_manager.console,
-        transient=False,  # We want to keep this visible
-        refresh_per_second=10,
-    ) as progress:
-        # Add port-forward task
-        task_id = progress.add_task(
-            description=f"{display_text} - Starting...", total=None
-        )
-
-        # Define the main async routine
-        async def main() -> None:
-            """Main async routine that runs port-forward and updates progress."""
-            proxy = None
-
-            try:
-                # Start proxy server if traffic monitoring is enabled
-                if use_proxy and proxy_port is not None:
-                    proxy = await start_proxy_server(
-                        local_port=int(local_port), target_port=proxy_port, stats=stats
-                    )
-
-                # Start the port-forward process
-                process = await run_port_forward()
-
-                # Start updating the progress display
-                progress_task = asyncio.create_task(
-                    update_progress(task_id, progress, process, proxy)
-                )
-
-                try:
-                    # Keep running until user interrupts with Ctrl+C
-                    await process.wait()
-
-                    # If we get here, the process completed or errored
-                    if process.returncode != 0:
-                        # Read error output
-                        stderr = await process.stderr.read() if process.stderr else b""
-                        error_msg = stderr.decode("utf-8").strip()
-                        stats.error_messages.append(error_msg)
-                        console_manager.print_error(f"Port-forward error: {error_msg}")
-
-                except asyncio.CancelledError:
-                    # User cancelled, terminate the process
-                    process.terminate()
-                    await process.wait()
-                    raise
-
-                finally:
-                    # Cancel the progress task
-                    if not progress_task.done():
-                        progress_task.cancel()
-                        with suppress(asyncio.CancelledError):
-                            await asyncio.wait_for(progress_task, timeout=0.5)
-
-            finally:
-                # Clean up proxy server if it was started
-                if proxy:
-                    await stop_proxy_server(proxy)
-
-        # Set up event loop and run the async code
-        created_new_loop = False
-        loop = None
-
-        try:
-            # Get or create an event loop
-            try:
-                loop = asyncio.get_event_loop()
-                if loop.is_running():
-                    loop = asyncio.new_event_loop()
-                    asyncio.set_event_loop(loop)
-                    created_new_loop = True
-            except RuntimeError:
-                loop = asyncio.new_event_loop()
-                asyncio.set_event_loop(loop)
-                created_new_loop = True
-
-            # Run the main coroutine
-            loop.run_until_complete(main())
-
-        except KeyboardInterrupt:
-            # Handle Ctrl+C gracefully
-            stats.current_status = "Cancelled (User)"
-            console_manager.print_note("\nPort-forward cancelled by user")
-            return Error(error="Port-forward cancelled by user")
-
-        except asyncio.CancelledError:
-            # Handle cancellation
-            stats.current_status = "Cancelled"
-            console_manager.print_note("\nPort-forward cancelled")
-            return Error(error="Port-forward cancelled")
-
-        except Exception as e:
-            # Handle other errors
-            stats.current_status = "Error"
-            stats.error_messages.append(str(e))
-            console_manager.print_error(f"\nPort-forward error: {e!s}")
-            return Error(error=f"Port-forward error: {e}", exception=e)
-
-        finally:
-            # Clean up
-            if created_new_loop and loop is not None:
-                loop.close()
-
-    # Calculate elapsed time
-    elapsed_time = time.time() - start_time
-
-    # Show final message with elapsed time
-    console_manager.print_note(
-        f"\n[bold]Port-forward session ended after "
-        f"[italic]{elapsed_time:.1f}s[/italic][/bold]"
+    # Call the worker function in live_display.py
+    return _execute_port_forward_with_live_display(
+        resource=resource,
+        args=args,
+        output_flags=output_flags,
+        port_mapping=port_mapping,
+        local_port=local_port,
+        remote_port=remote_port,
+        display_text=display_text,
+        summary_prompt_func=summary_prompt_func,
     )
-
-    # Create and display a table with connection statistics
-    table = Table(title=f"Port-forward {resource} Connection Summary")
-
-    # Add columns
-    table.add_column("Metric", style="cyan")
-    table.add_column("Value", style="green")
-
-    # Add rows with connection statistics
-    table.add_row("Status", stats.current_status)
-    table.add_row("Resource", resource)
-    table.add_row("Port Mapping", f"localhost:{local_port} → {remote_port}")
-    table.add_row("Duration", f"{elapsed_time:.1f}s")
-    table.add_row("Connected Time", f"{stats.elapsed_connected_time:.1f}s")
-    table.add_row("Connection Attempts", str(stats.connections_attempted))
-    table.add_row("Successful Connections", str(stats.successful_connections))
-
-    # Add proxy information if enabled
-    if stats.using_proxy:
-        table.add_row("Traffic Monitoring", "Enabled")
-        table.add_row("Proxy Mode", "Active")
-
-    # Add traffic information if available
-    if stats.traffic_monitoring_enabled:
-        table.add_row("Data Sent", f"{stats.bytes_sent} bytes")
-        table.add_row("Data Received", f"{stats.bytes_received} bytes")
-
-    # Add any error messages
-    if stats.error_messages:
-        table.add_row("Errors", "\n".join(stats.error_messages))
-
-    # Display the table
-    console_manager.console.print(table)
-
-    # Prepare forward info for memory
-    forward_info = f"Port-forward {resource} {port_mapping} ran for {elapsed_time:.1f}s"
-
-    # Create command string for memory
-    command_str = f"port-forward {resource} {' '.join(args)}"
-
-    # If vibe output is enabled, generate a summary using the LLM
-    vibe_output = ""
-    has_error = bool(stats.error_messages)
-
-    if output_flags.show_vibe:
-        try:
-            # Get the prompt function
-            summary_prompt_func = port_forward_prompt
-
-            # Get LLM summary of the port-forward session
-            model_adapter = get_model_adapter()
-            model = model_adapter.get_model(output_flags.model_name)
-
-            # Create detailed info for the prompt
-            detailed_info = {
-                "resource": resource,
-                "port_mapping": port_mapping,
-                "local_port": local_port,
-                "remote_port": remote_port,
-                "duration": f"{elapsed_time:.1f}s",
-                "command": command_str,
-                "status": stats.current_status,
-                "connected_time": f"{stats.elapsed_connected_time:.1f}s",
-                "connection_attempts": stats.connections_attempted,
-                "successful_connections": stats.successful_connections,
-                "traffic_monitoring_enabled": stats.traffic_monitoring_enabled,
-                "using_proxy": stats.using_proxy,
-                "bytes_sent": stats.bytes_sent,
-                "bytes_received": stats.bytes_received,
-                "errors": stats.error_messages,
-            }
-
-            # Format as YAML for the prompt
-            detailed_yaml = yaml.safe_dump(detailed_info, default_flow_style=False)
-
-            # Get the prompt template and format it
-            summary_prompt = summary_prompt_func()
-            prompt = summary_prompt.format(output=detailed_yaml, command=command_str)
-
-            # Execute the prompt to get a summary
-            vibe_output = model_adapter.execute(model, prompt)
-
-            # Display the vibe output
-            if vibe_output:
-                console_manager.print_vibe(vibe_output)
-
-        except Exception as e:
-            # Don't let errors in vibe generation break the command
-            console_manager.print_error(f"Error generating summary: {e}")
-            logger.error(f"Error generating port-forward summary: {e}", exc_info=True)
-
-    # Update memory with the port-forward information
-    update_memory(
-        command_str,
-        forward_info,
-        vibe_output,  # Now using the generated vibe output
-        output_flags.model_name,
-    )
-
-    # Return appropriate result
-    if has_error:
-        return Error(
-            error="\n".join(stats.error_messages)
-            or "Port-forward completed with errors",
-        )
-    else:
-        return Success(
-            message=(
-                f"Port-forward {resource} {port_mapping} completed "
-                f"successfully ({elapsed_time:.1f}s)"
-            ),
-            data=vibe_output,
-        )
